sensortools - Move code into lib for faster boot if disabled
