+++ apps/sensortools/metadata.json
@@ -12,6 +12,7 @@
   "storage": [
     {"name":"sensortools.0.boot.js","url":"boot.js"},
     {"name":"sensortools.settings.js","url":"settings.js"},
+    {"name":"sensortools","url":"lib.js"},
     {"name":"sensortools.default.json","url":"default.json"}
   ]
 }
